update: refactor for sdk
