+++ XUnit-BrowserStack/.config/dotnet-tools.json
@@ -1,0 +1,12 @@
+{
+  "version": 1,
+  "isRoot": true,
+  "tools": {
+    "browserstack-sdk": {
+      "version": "0.0.4",
+      "commands": [
+        "browserstack-sdk"
+      ]
+    }
+  }
+}
