--- XUnit-BrowserStack/config.json
@@ -1,29 +1,0 @@
-{
-  "server": "hub.browserstack.com",
-  "user": "BROWSERSTACK_USERNAME",
-  "key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "bstack:options": {
-      "buildName": "xunit-browserstack",
-      "sessionName": "test",
-      "debug": "true"
-    }
-  },
-
-  "environments": [
-    {
-      "browserName": "chrome"
-    },
-    {
-      "browserName": "firefox"
-    },
-    {
-      "browserName": "safari"
-    }
-  ],
-
-  "localOptions": {
-    "forceLocal": true
-  }
-}
